package: move functions:
- rename daautils -> utils
- move datautils.Vector -> utils/vectors.Vector
- rename ConvertCTypes -> Reinterpret cast
- move other wrapper stuff to internal

reaons: wrapper functions should not be exported for public use (some of them deffinitly should not be used externally)
